% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttm_i_and_n.R
\name{ttm_i}
\alias{ttm_i}
\title{Iteration count for ttm()}
\usage{
ttm_i()
}
\value{
int for ttm iteration, starting at 1
}
\description{
Iteration count for ttm()
}
\examples{
ttm(100, {
  x <- runif(1)
  cat('ttm i =', ttm_i(), 'n =', ttm_n(), 'x =', x, '\n')
  ttm_expect_true(x < 0.1)
})
}
