% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttm_i_and_n.R
\name{ttm_n}
\alias{ttm_n}
\title{Max number of attempts for ttm()}
\usage{
ttm_n()
}
\value{
int for ttm number of attempts
}
\description{
Max number of attempts for ttm()
}
\examples{
ttm(100, {
  x <- runif(1)
  cat('ttm i =', ttm_i(), 'n =', ttm_n(), 'x =', x, '\n')
  ttm_expect_true(x < 0.1)
})
}
