% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttm.R
\name{ttm}
\alias{ttm}
\title{Test that with multiple attempts}
\usage{
ttm(n, expr, verbose = 0)
}
\arguments{
\item{n}{Maximum number of attempts}

\item{expr}{Expression to evaluate}

\item{verbose}{Amount that should be printed}
}
\value{
Nothing
}
\description{
Test that with multiple attempts
}
\examples{

ttm(20, {
  xx <- runif(1)
  cat('xx', xx, "\n")
  ttm_expect_true(xx < .2)
})
}
